+++ server/helpers/dbImport/import.ts
@@ -1,6 +1,7 @@
 import { Writable } from "stream";
 import { Database, Statement } from "better-sqlite3";
 import { PoolClient } from "pg";
+import SQL from "sql-template-strings";
 import copy from "pg-copy-streams";
 import configs from "../configs";
 import { printProgress } from "./pretty";
@@ -11,16 +12,22 @@ export abstract class Import {
   abstract display: string;
   batchSizeMul = 1;
   initialKey: any[] = [-1];
+  useTemp = false;
   
-  abstract totalQuery: string;
-  abstract inputQuery: string;
-  abstract outputQuery: string;
+  abstract outputTable: string;
+  abstract totalQuery(): string;
+  abstract inputQuery(): string;
+  abstract outputQuery(table: string): string;
   
   private totalCount: number | null = null;
   
   total() {
-    if(this.totalCount === null) this.totalCount = this.hydrus.prepare(this.totalQuery).raw().get()[0];
+    if(this.totalCount === null) this.totalCount = this.hydrus.prepare(this.totalQuery()).raw().get()[0];
     return this.totalCount!;
+  }
+  
+  resetTotal(total: null | number = null) {
+    this.totalCount = total;
   }
   
   async importBatch(lastKey: any[], limit: number, input: Statement, output: Writable): Promise<any[] | null> {
@@ -40,7 +47,7 @@ export abstract class Import {
     printProgress(false, this.display);
     
     const batchSize = Math.ceil(configs.importBatchSize * this.batchSizeMul);
-    const total = await this.total();
+    const total = this.total();
     let count = 0;
     
     if(total === 0) {
@@ -50,10 +57,10 @@ export abstract class Import {
     
     printProgress([0, total], this.display);
     
-    await this.beforeImport();
+    const outputTable = await this.beforeImport();
     
-    const input = this.hydrus.prepare(this.inputQuery).raw(true);
-    const output: Writable = await this.postgres.query(copy.from(this.outputQuery));
+    const input = this.hydrus.prepare(this.inputQuery()).raw(true);
+    const output: Writable = await this.postgres.query(copy.from(this.outputQuery(outputTable)));
     
     let lastKey: any[] = this.initialKey;
     
@@ -79,6 +86,23 @@ export abstract class Import {
     printProgress([total, total], this.display);
   }
   
-  async beforeImport() {}
-  async afterImport() {}
+  async beforeImport() {
+    if(this.useTemp) {
+      await this.postgres.query(`
+        CREATE TEMP TABLE ${this.outputTable}_temp (LIKE ${this.outputTable});
+      `);
+      return `${this.outputTable}_temp`;
+    }
+    
+    return this.outputTable;
+  }
+  
+  async afterImport() {
+    if(this.useTemp) {
+      await this.postgres.query(`
+        INSERT INTO ${this.outputTable} SELECT * FROM ${this.outputTable}_temp ON CONFLICT DO NOTHING;
+        DROP TABLE ${this.outputTable}_temp;
+      `);
+    }
+  }
 }